Refactor Ava
Add documentation throughout, reorganize the codebase, delete dead code, and
mark incomplete, stubbed code as such.

+++ shared/datatypes/phone.go
@@ -1,6 +1,7 @@
 package dt
 
+// Phone represents a phone as a flexid from the database.
 type Phone struct {
-	Id     int
+	Id     uint64
 	Number string `db:"flexid"`
 }
